requirements.txt
python-vlc
youtube_dl
pafy
gtts
requests
